+++ cdpi_agent/cmd/extproc_agent/BUILD
@@ -30,7 +30,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//api/cdpi/v1alpha:cdpi_go_grpc",
-        "//cdpi_agent:agent",
+        "//cdpi_agent",
         "//cdpi_agent/enactment/extproc",
         "//cdpi_agent/internal/task",
         "//cdpi_agent/telemetry/extproc",
@@ -39,24 +39,18 @@ go_library(
         "@io_opentelemetry_go_contrib_instrumentation_google_golang_org_grpc_otelgrpc//:otelgrpc",
         "@io_opentelemetry_go_contrib_instrumentation_net_http_otelhttp//:otelhttp",
         "@io_opentelemetry_go_otel//attribute",
-        "@io_opentelemetry_go_otel//codes",
         "@io_opentelemetry_go_otel//propagation",
         "@io_opentelemetry_go_otel//semconv/v1.17.0:v1_17_0",
         "@io_opentelemetry_go_otel_exporters_zipkin//:zipkin",
         "@io_opentelemetry_go_otel_sdk//resource",
         "@io_opentelemetry_go_otel_sdk//trace",
-        "@io_opentelemetry_go_otel_trace//:trace",
-        "@org_golang_google_api//idtoken",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_grpc//backoff",
         "@org_golang_google_grpc//channelz/service",
         "@org_golang_google_grpc//codes",
         "@org_golang_google_grpc//credentials",
         "@org_golang_google_grpc//credentials/insecure",
-        "@org_golang_google_grpc//credentials/oauth",
         "@org_golang_google_grpc//status",
         "@org_golang_google_protobuf//proto",
-        "@org_golang_x_oauth2//:oauth2",
-        "@org_golang_x_sync//errgroup",
     ],
 )
